refactor: remove sharedStyles, move UI primitives to components/ui
