+++ src/components/common/Navbar.tsx
@@ -3,8 +3,8 @@ import { MdOutlineCollectionsBookmark } from "react-icons/md";
 import { Link, useNavigate } from "react-router-dom";
 import { useAppDispatch } from "../../hooks";
 import { createTree } from "../../redux/thunks";
-import { IconButtonMedium } from "../../styles/sharedStyles";
 import Settings from "../settings/Settings";
+import { IconButtonMedium } from "../ui/IconButton";
 import WelcomeInfoButton from "../WelcomeInfoButton";
 import { useActionBar } from "./ActionBarContext";
 import PantheonLogo from "./PantheonLogo";
