+++ src/components/common/ToggleButton.tsx
@@ -1,6 +1,5 @@
 import type React from "react";
 import { useState } from "react";
-import { ToggleKnob, ToggleSwitch } from "../../styles/sharedStyles";
 
 interface ToggleButtonProps {
 	initialState: boolean;
@@ -20,9 +19,22 @@ const ToggleButton: React.FC<ToggleButtonProps> = ({
 	};
 
 	return (
-		<ToggleSwitch toggled={isToggled} onClick={handleToggle}>
-			<ToggleKnob toggled={isToggled} />
-		</ToggleSwitch>
+		<div
+			className="w-[50px] h-[25px] rounded-[25px] relative cursor-pointer transition-[background-color] duration-300"
+			style={{ backgroundColor: isToggled ? "#4caf50" : "#ccc" }}
+			onClick={handleToggle}
+			onKeyDown={(e) => {
+				if (e.key === "Enter" || e.key === " ") handleToggle();
+			}}
+			role="switch"
+			aria-checked={isToggled}
+			tabIndex={0}
+		>
+			<div
+				className="w-[23px] h-[23px] bg-white rounded-full absolute top-px transition-[left] duration-300"
+				style={{ left: isToggled ? "26px" : "1px" }}
+			/>
+		</div>
 	);
 };
